+++ Questionable/Configuration.cs
@@ -119,6 +119,7 @@ internal sealed class Configuration : IPluginConfiguration
         public bool SkipCrystalTowerRaids { get; set; }
         public bool PreventQuestCompletion { get; set; }
         public bool AbandonQuestBeforeCompletion { get; set; }
+        public bool RemoveFromPriorityWhenAbandoned { get; set; }
         public bool ShowWindowOnStart { get; set; }
         public bool StartMinimized { get; set; }
         public bool OpenEditor { get; set; }