+++ Questionable/DalamudInitializer.cs
@@ -78,6 +78,7 @@ internal sealed class DalamudInitializer : IDisposable
         _toastGui.ErrorToast += OnErrorToast;
         _toastGui.QuestToast += OnQuestToast;
         _configuration.Advanced.AbandonQuestBeforeCompletion = false;
+        _configuration.Advanced.RemoveFromPriorityWhenAbandoned = false;
         if (_configuration.Advanced.StartMinimized)
             _questWindow.IsMinimized = true;