+++ CHANGELOG.md
@@ -1,7 +1,9 @@
+## Features
 - Troubleshooting button -alydev
   - Click the 🤝 button at the bottom of Questionable's main window to copy information to your clipboard, then paste into discord and describe your issue. Thank you for helping us help you help us all. -alydev
-- Fix quest priority auto removal -alydev
-  - Fix index error -alydev
 - Add current quest state to troubleshoot message -alydev
 - ARR intersocietal quests -alydev
+- Advanced tab: (prevent quest completion > abandon quest before completion > remove from prio quests) config chain -alydev
+## Bug fixes
+- Item use after combat: now paths to within interact range -alydev
 - Fix some solo duties targetting. - Kage